change int to double

+++ src/main/java/com/gbccccc/javafxdrawer/shape/model/Circle.java
@@ -5,5 +5,5 @@ import lombok.Data;
 
 @Data
 public class Circle {
-    private int radius = 0;
+    private double radius = 0;
 }
